Check valid pose in actionServer goal callback
Check that the pose of the detected object is not 0. If so it means that the detection has been discarded by the tracker because the model state confidence was not high enough. Right now only the edge matching confidence was being checked and that caused poses containing NaN to be published in some cases.